Able to read the API and create page system.
the user is able to click on next to get the next 12 item from the Api.

Problem: Usercart need to be able to read ApiReader, the item is store in ApiReader but you can not have both
div and raw data in the same time.

Some idea: Turn ApiReader to raw data and let Store_list print the data out in div.
Create another class in ApiReader to return raw data.